+++ src/deck/core/deckctrl_dfu_memory.c
